+++ website/src/site/ts/ssr.ts
@@ -1,3 +1,0 @@
-// this file imports only files that will be SSRd e.g. if you can't SSR a component, don't import it here.
-import './components/test-component.js';
-import './components/my-counter.js';